+++ sources/packages/backend/libs/test-utils/src/index.ts
@@ -1,6 +1,7 @@
-export * from "./designation-agreement";
-export * from "./institution";
-export * from "./institution-location";
-export * from "./user";
-export * from "./institution-user";
-export * from "./institution-user-auth";
+export * from "./factories/designation-agreement";
+export * from "./factories/institution";
+export * from "./factories/institution-location";
+export * from "./factories/user";
+export * from "./factories/institution-user";
+export * from "./factories/institution-user-auth";
+export * from "./constants";
